+++ docs/shared_combiners_catalog/rhel.rst
@@ -1,0 +1,3 @@
+.. automodule:: insights.combiners.rhel
+   :members:
+   :show-inheritance:
