--- docs/shared_components_catalog/rhel.rst
@@ -1,3 +1,0 @@
-.. automodule:: insights.components.rhel
-   :members:
-   :show-inheritance:
